+++ JtProject/src/main/java/com/jtspringproject/JtSpringProject/controller/AdminController.java
@@ -165,7 +165,7 @@ public class AdminController {
 	}
 
 	@RequestMapping(value = "products/add",method=RequestMethod.POST)
-	public String addProduct(@RequestParam("name") String name,@RequestParam("categoryid") int categoryId ,@RequestParam("price") int price,@RequestParam("weight") int weight, @RequestParam("quantity")int quantity,@RequestParam("description") String description,@RequestParam("productImage") String productImage) {
+	public String addProduct(@RequestParam("name") String name,@RequestParam("categoryid") int categoryId ,@RequestParam("price") double price,@RequestParam("weight") double weight, @RequestParam("quantity")int quantity,@RequestParam("description") String description,@RequestParam("productImage") String productImage) {
 		System.out.println(categoryId);
 		Category category = this.categoryService.getCategory(categoryId);
 		Product product = new Product();
@@ -192,10 +192,12 @@ public class AdminController {
 		mView.addObject("product", product);
 		return mView;
 	}
-	
-	@RequestMapping(value = "products/update/{id}",method=RequestMethod.POST)
-	public String updateProduct(@PathVariable("id") int id ,@RequestParam("name") String name,@RequestParam("categoryid") int categoryId ,@RequestParam("price") int price,@RequestParam("weight") int weight, @RequestParam("quantity")int quantity,@RequestParam("description") String description,@RequestParam("productImage") String productImage)
-	{
+
+	@RequestMapping(value = "products/update/{id}", method=RequestMethod.POST)
+	public String updateProduct(@PathVariable("id") int id, @RequestParam("name") String name,
+								@RequestParam("categoryid") int categoryId, @RequestParam("price") double price,
+								@RequestParam("weight") double weight, @RequestParam("quantity")int quantity,
+								@RequestParam("description") String description, @RequestParam("productImage") String productImage){
 
 //		this.productService.updateProduct();
 		return "redirect:/admin/products";
